make interact working
