+++ app.py
@@ -7,11 +7,9 @@ from models.db import db
 from utils.config import IS_DEV
 
 app = Flask(__name__, static_url_path='/static')
-
 app.secret_key = os.environ.get('FLASK_SECRET', '+%+3Q23!zbc+!Dd@')
 app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL')
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-
 db.init_app(app)
 
 cors = CORS(app,
