+++ utils/login_manager.py
@@ -1,42 +1,15 @@
+""" login_manager """
 from flask import current_app as app, jsonify
-from flask_login import LoginManager, login_user
+from flask_login import LoginManager
 
 from models.api_errors import ApiErrors
 from models.user import User
+from utils.credentials import get_user_with_credentials
 
 app.login_manager = LoginManager()
 app.login_manager.init_app(app)
-User = User
 
 app.config['REMEMBER_COOKIE_DURATION'] = 365 * 24 * 3600
-
-
-def get_user_with_credentials(identifier, password):
-    errors = ApiErrors()
-    errors.status_code = 401
-
-    if identifier is None:
-        errors.addError('identifier', 'Identifiant manquant')
-    if password is None:
-        errors.addError('password', 'Mot de passe manquant')
-    errors.maybeRaise()
-
-    user = User.query.filter_by(email=identifier).first()
-
-    if not user:
-        errors.addError('identifier', 'Identifiant incorrect')
-        raise errors
-    if not user.isValidated:
-        errors.addError('identifier', "Ce compte n'est pas validé.")
-        raise errors
-    if not user.checkPassword(password):
-        errors.addError('password', 'Mot de passe incorrect')
-        raise errors
-
-    login_user(user, remember=True)
-    return user
-
-app.get_user_with_credentials = get_user_with_credentials
 
 @app.login_manager.user_loader
 def get_user_with_id(user_id):
